sql: do not set is_temporary for ephemeral spaces
Up to this point, the is_temporary flag has been set in all
ephemeral spaces.

At some point, it became possible to use spaces created in Lua in
SQL statements. Since it is allowed to create temporary spaces in
Lua, not all temporary spaces in SQL are ephemeral. To separate
the temporary and ephemeral spaces, is_ephemeral flag is now set
in the ephemeral spaces and is_temporary flag is not set.

Close #4139